Adicionando confirmação de exclusão na página de edições e editando mensagem de exclusão

+++ src/pages/Gestao/GestaoVerificarEdicao/GestaoVerificarEdicao.tsx
@@ -18,6 +18,8 @@ import EditIcon from '@mui/icons-material/Edit'
 import { UserContext } from '../../../context/UserContex'
 import { Header } from '../../../components/Header/Header'
 import { CalendarioContext } from '../../../context/CalendarioContext'
+import { TOptionsConfirmDialog } from '../../../utils/interfaces'
+import { ConfirmDialog } from '../../../components/ConfirmDialog'
 
 export const GestaoVerificarEdicao = () => {
   const { edicao } = useParams()
@@ -35,6 +37,12 @@ export const GestaoVerificarEdicao = () => {
   useLayoutEffect(() => {
     getEtapas(idEdicao)
   }, [])
+
+  const [confirmDialog, setConfirmDialog] = React.useState<TOptionsConfirmDialog>({
+    isOpen: false,
+    title: "",
+    onConfirm: () => { }
+  });
 
   return (
     <>
@@ -112,15 +120,27 @@ export const GestaoVerificarEdicao = () => {
                       }}
                     />
                     <HighlightOffIcon
-                      id={`deletar-etapa-${index}`}
-                      onClick={() => {
-                        deleteEtapa(etapa.idEtapa, idEdicao)
-                      }}
-                      sx={{
+                      onClick={(event) => {
+                        setConfirmDialog({
+                          isOpen: true,
+                          title: `Confirma a exclusão da etapa ${etapa.nome}?`,
+                          onConfirm: () => {
+                            setConfirmDialog({
+                              ...confirmDialog,
+                              isOpen: false
+                            })
+                            deleteEtapa(etapa.idEtapa, idEdicao)
+                          }
+                        });
+                      }} sx={{
                         cursor: 'pointer',
-                        transition: '100ms all ease-in-out',
-                        '&:hover': { color: '#1e62fe' }
-                      }}
+                        width: '25px',
+                        height: '25px',
+                        "&:hover": { color: 'red', transform: 'scale(1.05)' },
+                        "& :active": {
+                          transform: 'scale(.99)',
+                        }
+                      }} 
                     />
                   </Box>
                 </Box>
@@ -217,21 +237,29 @@ export const GestaoVerificarEdicao = () => {
                                   }}
                                 />
                               </TableCell>
-
-                              <TableCell align="right" width={'40px'}>
+                              <TableCell align="right" width={'40px'}>   
                                 <HighlightOffIcon
-                                  id={`deletar-processo-${procIndex}`}
-                                  onClick={() =>
-                                    deleteProcesso(
-                                      processo.idProcesso,
-                                      idEdicao
-                                    )
-                                  }
-                                  sx={{
+                                  onClick={(event) => {
+                                    setConfirmDialog({
+                                      isOpen: true,
+                                      title: `Confirma a exclusão do processo ${processo.nome}?`,
+                                      onConfirm: () => {
+                                        setConfirmDialog({
+                                          ...confirmDialog,
+                                          isOpen: false
+                                        })
+                                        deleteProcesso(processo.idProcesso, idEdicao)
+                                      }
+                                    });
+                                  }} sx={{
                                     cursor: 'pointer',
-                                    transition: '100ms all ease-in-out',
-                                    '&:hover': { color: '#1e62fe' }
-                                  }}
+                                    width: '25px',
+                                    height: '25px',
+                                    "&:hover": { color: 'red', transform: 'scale(1.05)' },
+                                    "& :active": {
+                                      transform: 'scale(.99)',
+                                    }
+                                  }} 
                                 />
                               </TableCell>
                             </TableRow>
@@ -240,6 +268,10 @@ export const GestaoVerificarEdicao = () => {
                       )}
                     </TableBody>
                   </Table>
+                  <ConfirmDialog
+                    confirmDialog={confirmDialog}
+                    setConfirmDialog={setConfirmDialog}
+                  />
                 </TableContainer>
               </Box>
             </Box>
